+++ src/activitypub/builders/mod.rs
@@ -8,6 +8,7 @@ pub mod delete_person;
 pub mod follow;
 pub mod like;
 pub mod move_person;
+pub mod offer_agreement;
 pub mod reject_follow;
 pub mod remove_note;
 pub mod remove_person;